Rename classes. Restore legacy post_selection path. Move new implementation to bitflip_checks sub-package.
