+++ docs/apidocs/index.rst
@@ -10,6 +10,8 @@ This package contains functionality which is meant to supplement workflows invol
    qiskit_addon_utils.coloring
    qiskit_addon_utils.noise_management
    qiskit_addon_utils.noise_management.post_selection.transpiler.passes
+   qiskit_addon_utils.noise_management.bitflip_checks
+   qiskit_addon_utils.noise_management.bitflip_checks.transpiler.passes
    qiskit_addon_utils.problem_generators
    qiskit_addon_utils.slicing
    qiskit_addon_utils.slicing.transpiler.passes
