+++ docs/apidocs/qiskit_addon_utils.noise_management.bitflip_checks.rst
@@ -1,0 +1,14 @@
+=============================================================================
+bitflip_checks (:mod:`qiskit_addon_utils.noise_management.bitflip_checks`)
+=============================================================================
+
+.. automodule:: qiskit_addon_utils.noise_management.bitflip_checks
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
+
+.. currentmodule:: qiskit_addon_utils.noise_management.bitflip_checks
+
+.. autoclass:: PostSelectionSummary
+.. autoclass:: PostSelector
+.. autoclass:: PostSelectionStrategy
