+++ docs/apidocs/qiskit_addon_utils.noise_management.bitflip_checks.transpiler.passes.rst
@@ -1,0 +1,20 @@
+==============================================================================================================
+bitflip_checks_transpiler_passes (:mod:`qiskit_addon_utils.noise_management.bitflip_checks.transpiler.passes`)
+==============================================================================================================
+
+.. automodule:: qiskit_addon_utils.noise_management.bitflip_checks.transpiler.passes
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
+
+.. currentmodule:: qiskit_addon_utils.noise_management.bitflip_checks.transpiler.passes
+
+.. autosummary::
+   :toctree: ../stubs/
+   :nosignatures:
+
+   AddPostCircuitBitFlipChecks
+   AddPreCircuitBitFlipChecks
+   AddSpectatorPostCircuitBitFlipChecks
+   AddSpectatorPreCircuitBitFlipChecks
+   XSlowGate
