+++ docs/apidocs/qiskit_addon_utils.noise_management.post_selection.transpiler.passes.rst
@@ -16,5 +16,3 @@ post_selection_transpiler_passes (:mod:`qiskit_addon_utils.noise_management.post
    AddSpectatorMeasures
    AddPostSelectionMeasures
    XSlowGate
-   AddPreSelectionMeasures
-   AddSpectatorMeasuresPreSelection
