+++ docs/apidocs/qiskit_addon_utils.noise_management.rst
@@ -13,7 +13,3 @@ noise_management (:mod:`qiskit_addon_utils.noise_management`)
 .. autofunction:: gamma_from_noisy_boxes
 .. autoclass:: PostSelectionSummary
 .. autoclass:: PostSelector
-
-.. currentmodule:: qiskit_addon_utils.noise_management.post_selection
-
-.. autoclass:: PostSelectionStrategy
