+++ qiskit_addon_utils/noise_management/__init__.py
@@ -15,10 +15,7 @@
 """Noise management tools."""
 
 from .gamma_factor import gamma_from_noisy_boxes
-from .post_selection import (
-    PostSelectionSummary,
-    PostSelector,
-)
+from .post_selection import PostSelectionSummary, PostSelector
 from .trex_factors import trex_factors
 
 __all__ = [
