+++ qiskit_addon_utils/noise_management/bitflip_checks/transpiler/passes/utils.py
@@ -26,6 +26,8 @@ def validate_op_is_supported(node: DAGOpNode):
         * Standard gates.
         * Barriers.
         * Measurements.
+        * Resets.
+        * Delays.
         * Control-flow operations.
 
     Args:
@@ -36,7 +38,7 @@ def validate_op_is_supported(node: DAGOpNode):
     """
     if (
         node.is_standard_gate()
-        or node.op.name in ["barrier", "measure"]
+        or node.op.name in ["barrier", "measure", "reset", "xslow", "delay"]
         or isinstance(node.op, ControlFlowOp)
     ):
         return
