+++ releasenotes/notes/bitflip-checks-6bc8f8a28023f8c9.yaml
@@ -1,0 +1,58 @@
+---
+features:
+  - |
+    Added a new :mod:`qiskit_addon_utils.noise_management.bitflip_checks` sub-package for
+    inserting bit-flip checks into a circuit. It supports both *post-selection* (filtering
+    shots based on parity checks at the end of the circuit) and *pre-selection* (filtering
+    out shots where one or more qubits were not properly initialized to the ground state at
+    the start of the circuit). The two techniques can be used independently or combined.
+
+    New transpiler passes in
+    :mod:`qiskit_addon_utils.noise_management.bitflip_checks.transpiler.passes`:
+
+      * :class:`~qiskit_addon_utils.noise_management.bitflip_checks.transpiler.passes.AddPostCircuitBitFlipChecks`
+        adds a narrowband X-pulse (an ``xslow`` gate, or an ``rx``-based pulse) sandwiched
+        between two measurements at the *end* of the circuit for every active qubit with a
+        terminal measurement. A parity violation between the two readings indicates an
+        unwanted bit flip and the shot is discarded.
+      * :class:`~qiskit_addon_utils.noise_management.bitflip_checks.transpiler.passes.AddPreCircuitBitFlipChecks`
+        adds the analogous narrowband X-pulse, :class:`~qiskit.circuit.library.XGate`, and
+        measurement at the *beginning* of the circuit for every active qubit with a terminal
+        measurement. Results that read ``1`` indicate a bad initialization and are discarded.
+      * :class:`~qiskit_addon_utils.noise_management.bitflip_checks.transpiler.passes.AddSpectatorPostCircuitBitFlipChecks`
+        extends the post-selection parity check to spectator qubits (and, by default, to
+        active but unterminated qubits).
+      * :class:`~qiskit_addon_utils.noise_management.bitflip_checks.transpiler.passes.AddSpectatorPreCircuitBitFlipChecks`
+        extends the pre-selection check to spectator qubits, writing to a register named
+        ``spec_pre``.
+  - |
+    Added :class:`~qiskit_addon_utils.noise_management.bitflip_checks.PostSelector` and
+    :class:`~qiskit_addon_utils.noise_management.bitflip_checks.PostSelectionSummary` for
+    computing shot-selection masks from circuits processed by the ``bitflip_checks`` passes.
+    :meth:`~qiskit_addon_utils.noise_management.bitflip_checks.PostSelector.compute_mask`
+    accepts a ``mode`` keyword argument with values ``"post"`` (the default), ``"pre"``, or
+    ``"both"``. ``"pre"`` keeps only shots that passed pre-selection, ``"post"`` keeps only
+    shots that passed post-selection, and ``"both"`` combines the two masks with a logical
+    AND. ``compute_mask`` raises ``ValueError`` if the requested mode is not available on the
+    circuit (for example, ``mode="pre"`` on a circuit with no pre-selection measurements).
+  - |
+    :meth:`~qiskit_addon_utils.noise_management.bitflip_checks.PostSelector.from_circuit` and
+    :meth:`~qiskit_addon_utils.noise_management.bitflip_checks.PostSelectionSummary.from_circuit`
+    accept ``pre_selection_suffix`` and ``spectator_cregs`` keyword arguments.
+    ``spectator_cregs`` labels which primary registers hold spectator measurements
+    (defaulting to ``["spec"]`` to match
+    :class:`~qiskit_addon_utils.noise_management.bitflip_checks.transpiler.passes.AddSpectatorPostCircuitBitFlipChecks`);
+    downstream code that wants only data-measurement registers can compute
+    ``summary.primary_cregs - summary.spectator_cregs``.
+    :class:`~qiskit_addon_utils.noise_management.bitflip_checks.PostSelectionSummary` exposes
+    the ``measure_map_ps``, ``measure_map_pre``, ``pre_selection_suffix``, and
+    ``spectator_cregs`` properties for the separate post- and pre-selection measurement maps
+    and related metadata.
+  - |
+    :class:`~qiskit_addon_utils.noise_management.bitflip_checks.PostSelectionStrategy`
+    selects between the ``"node"`` and ``"edge"`` parity-check strategies, which apply
+    uniformly to pre-selection, post-selection, and combined modes.
+  - |
+    Added the module-level constants ``DEFAULT_PRE_SELECTION_SUFFIX`` (``"_pre"``) and
+    ``DEFAULT_SPECTATOR_PRE_CREG_NAME`` (``"spec_pre"``) to
+    :mod:`qiskit_addon_utils.noise_management.constants`.
