--- releasenotes/notes/preselection-6bc8f8a28023f8c9.yaml
@@ -1,92 +1,0 @@
----
-features:
-  - |
-    Added pre-selection support to the :mod:`qiskit_addon_utils.noise_management.post_selection`
-    sub-package. Pre-selection filters out shots where one or more qubits were not properly
-    initialized to the ground state at the start of the circuit, complementing post-selection
-    (which filters shots based on parity checks at the end of the circuit). The two techniques
-    can be used independently or combined.
-
-    New transpiler passes:
-
-      * :class:`~qiskit_addon_utils.noise_management.post_selection.transpiler.passes.AddPreSelectionMeasures`
-        adds an ``xslow`` (or ``rx``-based) narrowband X-pulse, followed by an :class:`~qiskit.circuit.library.XGate`
-        and a measurement, at the *beginning* of the circuit for every active qubit with a terminal measurement.
-        Results that read ``1`` indicate a bad initialization and are discarded.
-      * :class:`~qiskit_addon_utils.noise_management.post_selection.transpiler.passes.AddSpectatorMeasuresPreSelection`
-        extends the pre-selection check to spectator qubits (and, by default, to active but
-        unterminated qubits), writing to a register named ``spec_pre``.
-
-    New module-level constants in :mod:`qiskit_addon_utils.noise_management.constants`:
-    ``DEFAULT_PRE_SELECTION_SUFFIX`` (``"_pre"``) and ``DEFAULT_SPECTATOR_PRE_CREG_NAME``
-    (``"spec_pre"``).
-  - |
-    :meth:`~qiskit_addon_utils.noise_management.PostSelector.compute_mask` accepts a new
-    ``mode`` keyword argument with values ``"post"`` (the default, unchanged behavior),
-    ``"pre"``, or ``"both"``. ``"pre"`` keeps only shots that passed pre-selection,
-    ``"post"`` keeps only shots that passed post-selection, and ``"both"`` combines the two
-    masks with a logical AND. ``compute_mask`` raises ``ValueError`` if the requested mode
-    is not available on the circuit (for example, ``mode="pre"`` on a circuit that contains
-    no pre-selection measurements).
-  - |
-    :meth:`~qiskit_addon_utils.noise_management.PostSelector.from_circuit` and
-    :meth:`~qiskit_addon_utils.noise_management.PostSelectionSummary.from_circuit` accept new
-    ``pre_selection_suffix`` and ``spectator_cregs`` keyword arguments. ``spectator_cregs``
-    labels which primary registers hold spectator measurements (defaulting to ``["spec"]`` to
-    match :class:`~qiskit_addon_utils.noise_management.post_selection.transpiler.passes.AddSpectatorMeasures`);
-    downstream code that wants only data-measurement registers can compute
-    ``summary.primary_cregs - summary.spectator_cregs``.
-  - |
-    :class:`~qiskit_addon_utils.noise_management.PostSelectionSummary` gained the
-    ``measure_map_ps``, ``measure_map_pre``, ``pre_selection_suffix``, and
-    ``spectator_cregs`` properties, exposing the separate post- and pre-selection
-    measurement maps and related metadata.
-  - |
-    :class:`~qiskit_addon_utils.noise_management.post_selection.transpiler.passes.AddPostSelectionMeasures`
-    accepts a new ``ignore_creg_suffixes`` keyword argument (defaulting to ``["_pre"]``)
-    so that pre-selection registers are not duplicated when adding post-selection. The pass
-    also skips ``reset`` operations, allowing it to be applied to circuits that already
-    contain pre-selection structure.
-  - |
-    :class:`~qiskit_addon_utils.noise_management.post_selection.transpiler.passes.AddSpectatorMeasures`
-    gained ``x_pulse_type``, ``ignore_creg_suffixes``, and ``post_selection_suffix``
-    keyword arguments. When a circuit already contains the barrier sandwich produced by
-    :class:`~qiskit_addon_utils.noise_management.post_selection.transpiler.passes.AddPostSelectionMeasures`,
-    the pass now splices the spectator parity check into that structure so that the
-    spectator pulses run alongside the data-qubit pulses.
-  - |
-    :class:`~qiskit_addon_utils.noise_management.post_selection.PostSelectionStrategy`
-    is now documented in the API reference. It is unchanged in behavior — ``"node"``
-    and ``"edge"`` strategies apply uniformly to pre-selection, post-selection, and
-    combined modes.
-upgrade:
-  - |
-    :class:`~qiskit_addon_utils.noise_management.post_selection.transpiler.passes.AddSpectatorMeasures`
-    now performs a full parity check on each spectator qubit — a measurement, a narrowband
-    X-pulse (an ``xslow`` gate, or 20 ``rx(pi/20)`` gates), and a second measurement — and
-    writes the two readings to *two* classical registers (default names ``spec`` and
-    ``spec_ps``) rather than a single ``spec`` register. Downstream code that consumed
-    the former single-measurement ``spec`` register, or that relied on the previous
-    "just measure once" behavior, must be updated. :class:`~qiskit_addon_utils.noise_management.PostSelector`
-    handles the new structure automatically.
-  - |
-    The ``add_barrier`` keyword argument of
-    :class:`~qiskit_addon_utils.noise_management.post_selection.transpiler.passes.AddSpectatorMeasures`
-    has been removed. The pass now manages its own barriers, and when post-selection is
-    present it shares barriers with :class:`~qiskit_addon_utils.noise_management.post_selection.transpiler.passes.AddPostSelectionMeasures`
-    rather than emitting its own.
-  - |
-    :class:`~qiskit_addon_utils.noise_management.post_selection.transpiler.passes.AddSpectatorMeasures`
-    gained a new positional argument, ``x_pulse_type``, immediately after ``coupling_map``.
-    Callers that pass ``coupling_map`` positionally and all other arguments by keyword
-    are unaffected; callers that previously passed ``include_unmeasured`` or
-    ``spectator_creg_name`` positionally must update to keyword form.
-  - |
-    :class:`~qiskit_addon_utils.noise_management.PostSelectionSummary` no longer requires
-    that every primary register have a matching ``_ps`` register. A primary register may
-    now be paired with a pre-selection register, a post-selection register, or both, which
-    is what enables pre-only and pre+post workflows. Construction from circuits that
-    *do* contain post-selection registers is still validated strictly: every primary
-    register must have a matching ``_ps`` register of the same size. Pre-selection
-    validation is intrinsically lenient — only the subset of qubits that are
-    pre-selected is checked.
