+++ test/noise_management/bitflip_checks/passes/__init__.py
@@ -9,12 +9,3 @@
 # Any modifications or derivative works of this code must retain this
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
-
-# Reminder: update the RST file in docs/apidocs when adding new interfaces.
-"""A submodule with transpilation passes for post selection."""
-
-from .add_post_selection_measures import AddPostSelectionMeasures, XPulseType
-from .add_spectator_measures import AddSpectatorMeasures
-from .xslow_gate import XSlowGate
-
-__all__ = ["AddPostSelectionMeasures", "AddSpectatorMeasures", "XPulseType", "XSlowGate"]
