+++ test/noise_management/bitflip_checks/passes/test_node_passes.py
@@ -25,9 +25,9 @@ from __future__ import annotations
 import pytest
 from qiskit.circuit import QuantumCircuit
 from qiskit.transpiler import PassManager
-from qiskit_addon_utils.noise_management.post_selection.transpiler.passes import (
-    AddPostSelectionMeasures,
-    AddPreSelectionMeasures,
+from qiskit_addon_utils.noise_management.bitflip_checks.transpiler.passes import (
+    AddPostCircuitBitFlipChecks,
+    AddPreCircuitBitFlipChecks,
 )
 
 
@@ -66,7 +66,7 @@ def _meas_registers(circuit: QuantumCircuit, qubit_idx: int) -> list[str]:
 
 def test_post_sel_default():
     """Post-selection alone with default suffix: adds ``c_ps``."""
-    pm = PassManager([AddPostSelectionMeasures(x_pulse_type="rx")])
+    pm = PassManager([AddPostCircuitBitFlipChecks(x_pulse_type="rx")])
     result = pm.run(_make_circuit())
 
     assert _creg_map(result) == {"c": 3, "c_ps": 3}
@@ -76,7 +76,9 @@ def test_post_sel_default():
 
 def test_post_sel_custom_suffix():
     """Post-selection with custom suffix: adds ``c_check``, never ``c_ps``."""
-    pm = PassManager([AddPostSelectionMeasures(x_pulse_type="rx", post_selection_suffix="_check")])
+    pm = PassManager(
+        [AddPostCircuitBitFlipChecks(x_pulse_type="rx", post_selection_suffix="_check")]
+    )
     result = pm.run(_make_circuit())
 
     assert _creg_map(result) == {"c": 3, "c_check": 3}
@@ -86,7 +88,7 @@ def test_post_sel_custom_suffix():
 
 def test_pre_sel_default():
     """Pre-selection alone with default suffix: adds ``c_pre``."""
-    pm = PassManager([AddPreSelectionMeasures(x_pulse_type="rx")])
+    pm = PassManager([AddPreCircuitBitFlipChecks(x_pulse_type="rx")])
     result = pm.run(_make_circuit())
 
     assert _creg_map(result) == {"c": 3, "c_pre": 3}
@@ -96,7 +98,7 @@ def test_pre_sel_default():
 
 def test_pre_sel_custom_suffix():
     """Pre-selection with custom suffix: adds ``c_init``, never ``c_pre``."""
-    pm = PassManager([AddPreSelectionMeasures(x_pulse_type="rx", pre_selection_suffix="_init")])
+    pm = PassManager([AddPreCircuitBitFlipChecks(x_pulse_type="rx", pre_selection_suffix="_init")])
     result = pm.run(_make_circuit())
 
     assert _creg_map(result) == {"c": 3, "c_init": 3}
@@ -113,8 +115,8 @@ def test_pre_then_post_default():
     """Pre then post (default suffixes). No cross-contamination of registers."""
     pm = PassManager(
         [
-            AddPreSelectionMeasures(x_pulse_type="rx"),
-            AddPostSelectionMeasures(x_pulse_type="rx"),
+            AddPreCircuitBitFlipChecks(x_pulse_type="rx"),
+            AddPostCircuitBitFlipChecks(x_pulse_type="rx"),
         ]
     )
     result = pm.run(_make_circuit())
@@ -130,8 +132,8 @@ def test_post_then_pre_default():
     """Post then pre (default suffixes). Final structure matches ``pre_then_post``."""
     pm = PassManager(
         [
-            AddPostSelectionMeasures(x_pulse_type="rx"),
-            AddPreSelectionMeasures(x_pulse_type="rx"),
+            AddPostCircuitBitFlipChecks(x_pulse_type="rx"),
+            AddPreCircuitBitFlipChecks(x_pulse_type="rx"),
         ]
     )
     result = pm.run(_make_circuit())
@@ -147,8 +149,8 @@ def test_pre_then_post_custom():
     """Pre then post with custom suffixes; pre must be ignored by post."""
     pm = PassManager(
         [
-            AddPreSelectionMeasures(x_pulse_type="rx", pre_selection_suffix="_init"),
-            AddPostSelectionMeasures(
+            AddPreCircuitBitFlipChecks(x_pulse_type="rx", pre_selection_suffix="_init"),
+            AddPostCircuitBitFlipChecks(
                 x_pulse_type="rx",
                 post_selection_suffix="_check",
                 ignore_creg_suffixes=["_init"],
@@ -168,8 +170,8 @@ def test_post_then_pre_custom():
     """Post then pre with custom suffixes; post must be ignored by pre."""
     pm = PassManager(
         [
-            AddPostSelectionMeasures(x_pulse_type="rx", post_selection_suffix="_check"),
-            AddPreSelectionMeasures(
+            AddPostCircuitBitFlipChecks(x_pulse_type="rx", post_selection_suffix="_check"),
+            AddPreCircuitBitFlipChecks(
                 x_pulse_type="rx",
                 pre_selection_suffix="_init",
                 ignore_creg_suffixes=["_check"],
@@ -191,36 +193,36 @@ def test_post_then_pre_custom():
 
 
 def test_empty_circuit_post_selection():
-    """Empty circuit is unchanged by ``AddPostSelectionMeasures``."""
+    """Empty circuit is unchanged by ``AddPostCircuitBitFlipChecks``."""
     qc = QuantumCircuit(1)
-    pm = PassManager([AddPostSelectionMeasures()])
+    pm = PassManager([AddPostCircuitBitFlipChecks()])
     assert pm.run(qc) == qc
 
 
 def test_empty_circuit_pre_selection():
-    """Empty circuit is unchanged by ``AddPreSelectionMeasures``."""
+    """Empty circuit is unchanged by ``AddPreCircuitBitFlipChecks``."""
     qc = QuantumCircuit(1)
-    pm = PassManager([AddPreSelectionMeasures()])
+    pm = PassManager([AddPreCircuitBitFlipChecks()])
     assert pm.run(qc) == qc
 
 
 def test_invalid_x_pulse_type_post():
     """Unknown ``x_pulse_type`` is rejected by post-selection pass."""
     with pytest.raises(ValueError):
-        AddPostSelectionMeasures(x_pulse_type="rz")
+        AddPostCircuitBitFlipChecks(x_pulse_type="rz")
 
 
 def test_invalid_x_pulse_type_pre():
     """Unknown ``x_pulse_type`` is rejected by pre-selection pass."""
     with pytest.raises(ValueError):
-        AddPreSelectionMeasures(x_pulse_type="rz")
+        AddPreCircuitBitFlipChecks(x_pulse_type="rz")
 
 
 def test_pre_selection_no_measurements_returns_unchanged():
     """Active qubits but no measurements: pass exits early without modification."""
     qc = QuantumCircuit(1)
     qc.h(0)
-    pm = PassManager([AddPreSelectionMeasures()])
+    pm = PassManager([AddPreCircuitBitFlipChecks()])
     assert pm.run(qc) == qc
 
 
@@ -232,7 +234,7 @@ def test_pre_selection_only_ignored_registers_returns_unchanged():
     qc.h(0)
     qc.measure(0, 0)
     # Default ``ignore_creg_names=["spec"]`` filters out the only register.
-    pm = PassManager([AddPreSelectionMeasures()])
+    pm = PassManager([AddPreCircuitBitFlipChecks()])
     assert pm.run(qc) == qc
 
 
@@ -244,7 +246,7 @@ def test_unsupported_op_raises():
     qc = QuantumCircuit(1, 1)
     qc.append(Initialize("0"), [0])
     qc.measure(0, 0)
-    pm = PassManager([AddPostSelectionMeasures()])
+    pm = PassManager([AddPostCircuitBitFlipChecks()])
     with pytest.raises(TranspilerError, match="not supported"):
         pm.run(qc)
 
@@ -256,7 +258,7 @@ def test_delay_is_supported():
     qc = QuantumCircuit(1, 1)
     qc.append(Delay(100), [0])
     qc.measure(0, 0)
-    pm = PassManager([AddPostSelectionMeasures()])
+    pm = PassManager([AddPostCircuitBitFlipChecks()])
     # Should run without raising.
     pm.run(qc)
 
@@ -276,6 +278,6 @@ def test_post_selection_skips_user_reset():
     qc.reset(0)
     qc.measure(0, creg[0])
 
-    result = PassManager([AddPostSelectionMeasures()]).run(qc)
+    result = PassManager([AddPostCircuitBitFlipChecks()]).run(qc)
     # ``c_ps`` register added; qubit is still recognised as terminated despite the reset.
     assert {creg.name for creg in result.cregs} == {"c", "c_ps"}
